backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-multipart==0.0.9
openai==1.51.0
chromadb==0.5.5
httpx>=0.27.0,<0.28.0
python-dotenv==1.0.1
pdfplumber==0.11.4
python-docx==1.1.2
pydantic>=2.0
markdown>=3.0
